7-main.c

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -7,7 +7,7 @@
 
 int print_last_digit(int c)
 {
-	long long int b;
+	long int b;
 
 	if (c < 0)
 	{
